Refactor Multi-Agent Tracking Module and Introduce Quadratic Cost Utilities
- Split the MultiAgentTracking.jl file into multiple files for better organization:
  - Targets.jl: Contains the BobbingTarget struct and trajectory generation function.
  - TrackingProblems.jl: Defines TrackingEdge and TrackingProblem structs along with sheaf construction and trajectory extraction functions.
  - ScenarioResults.jl: Defines the ScenarioResult struct and its display methods.
  - QuadraticCosts.jl: Introduces utilities for building and solving quadratic control costs on the tracking sheaf.

- Implemented build_control_cost_matrix to create a global quadratic cost matrix for agent controls.
- Added solve_quadratic_on_basis to optimize control inputs while respecting Laplacian constraints.
- Updated run_scenario to incorporate optional quadratic cost handling.
- Added unit tests for quadratic cost utilities to ensure correctness of cost matrix construction and optimization.
